Remove Admin namespace
